+++ src/builder/index.ts
@@ -15,21 +15,20 @@ import * as rdp from '../other/rdp';
 import * as dds from '../other/dds';
 
 import * as xmltest from '../other/xmltest';
-import * as utils from '../other/utils';
 
 export class ModBuilder {
   _converters: { [index: string]: Converter } = {};
   _logger;
   _asAbsolutePath;
-  _variableAnnoMods;
+  _variables;
 
-  public constructor(logger: ILogger, asAbsolutePath: (relative: string) => string, variableAnnoMods: string) {
+  public constructor(logger: ILogger, asAbsolutePath: (relative: string) => string, variables: { [index: string]: string }) {
     rdp.init(asAbsolutePath('./external/'));
     dds.init(asAbsolutePath('./external/'));
 
     this._logger = logger;
     this._asAbsolutePath = asAbsolutePath;
-    this._variableAnnoMods = variableAnnoMods;
+    this._variables = variables;
 
     this._addConverter(new StaticConverter());
     this._addConverter(new Cf7Converter());
@@ -77,7 +76,8 @@ export class ModBuilder {
           cache,
           ci,
           modJson,
-          converterOptions: entry
+          converterOptions: entry,
+          variables: this._variables
         });
         if (!result) {
           return false;
@@ -107,7 +107,9 @@ export class ModBuilder {
   private _getOutFolder(filePath: string, modJson: any) {
     let outFolder = modJson.out;
     outFolder = outFolder.replace('${modName}', this._getModName(filePath, modJson));
-    outFolder = path.normalize(outFolder.replace('${annoMods}', this._variableAnnoMods));
+    if (this._variables['annoMods']) {
+      outFolder = path.normalize(outFolder.replace('${annoMods}', this._variables['annoMods']));
+    }
     if (!path.isAbsolute(outFolder)) {
       outFolder = path.join(path.dirname(filePath), outFolder);
     }